+++ cmd/angry-box/main.go
@@ -19,7 +19,7 @@ import (
 )
 
 var (
-	version = "v0.6.0"
+	version = "v0.6.1"
 	commit  = "none"
 	date    = "unknown"